// src/app/Models/UserDataModel.ts
// -----------------------------------------------------------------------
// <copyright file="UserDataModel.ts" company="Soloplan GmbH">
//     Copyright (c) Soloplan GmbH. All rights reserved.
// </copyright>
// -----------------------------------------------------------------------
export interface UserDataModel {
    Name?: string;
    FullName?: string;
    Organization?: string;
    Rememberme?: boolean;
}